Added a relativeDate implementation that is needed now that we no longer use FormatterKit.
instead of using the TTTTimeIntervalFormatter for dates that are less than a week ago we use our own relativeStringDate()
This implementation does not include future dates formatting ("from now"). It seems to me like "from now" was not used in the app (looked at scheduled posts)
let me know if I'm wrong and it is used and needed.